Allow services breakdown to span more than one page
Prior to this commit, if the services breakdown table could
not be fit on one page, it would be included on a separate page
after all other content.

This commit changes this to allow the services breakdown
to remain in the indented location and simply wrap onto the next page.